Se saco funcionalidad en la capa controller hasta PaymentController inclusive

+++ src/main/java/ar/edu/unnoba/poo2019/webapp/service/EventServiceImp.java
@@ -8,6 +8,7 @@ package ar.edu.unnoba.poo2019.webapp.service;
 import ar.edu.unnoba.poo2019.webapp.model.Event;
 import ar.edu.unnoba.poo2019.webapp.repository.EventRepository;
 import java.util.List;
+import java.util.Objects;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -16,18 +17,18 @@ import org.springframework.stereotype.Service;
  * @author root
  */
 @Service
-public class EventServiceImp implements EventService{
-    
+public class EventServiceImp implements EventService {
+
     @Autowired
     private SessionService sessionService;
-    
+
     @Autowired
     private EventRepository eventRepository;
-    
+
     @Override
     public List<Event> events() {
         return eventRepository.findAllEvents();
-        
+
     }
 
     @Override
@@ -42,27 +43,40 @@ public class EventServiceImp implements EventService{
     }
 
     @Override
-    public Event update(Long id, Event event) {
-        Event e = eventRepository.findById(id).get();
-        e.setEventName(event.getEventName());
-        e.setCapacity(event.getCapacity());
-        e.setCost(event.getCost());
-        e.setStartRegistrations(event.getStartRegistrations());
-        e.setEndRegistrations(event.getEndRegistrations());
-        e.setEventDate(event.getEventDate());
-        e.setPrivateEvent(event.isPrivateEvent());
-        e.setLugar(event.getLugar());
-        return eventRepository.save(e);  
+    public Event update(Long id, Event event) throws Exception {
+        Event oldEvent = this.find(id);
+        if (event.getCapacity() >= oldEvent.cantidadRegistrations()) {     // Controlo que la capacidad no sea menor a los inscriptos
+            if (oldEvent.hasRegistrations() && oldEvent.getCost() != event.getCost()) {   // Controlo que no se pueda cambiar el precio si hay registraciones
+                throw new Exception("No se puede cambiar el costo si ya hay inscriptos");
+            } else {
+                oldEvent.setEventName(event.getEventName());    // Guardo el evento
+                oldEvent.setCapacity(event.getCapacity());
+                oldEvent.setCost(event.getCost());
+                oldEvent.setStartRegistrations(event.getStartRegistrations());
+                oldEvent.setEndRegistrations(event.getEndRegistrations());
+                oldEvent.setEventDate(event.getEventDate());
+                oldEvent.setPrivateEvent(event.isPrivateEvent());
+                oldEvent.setLugar(event.getLugar());
+                return eventRepository.save(oldEvent);
+            }
+        }else{
+            throw new Exception("La capacidad no puede ser menor que la cantidad de inscriptos");
+        }
     }
 
     @Override
-    public void delete(Long id) {
-        eventRepository.deleteById(id);    
+    public void delete(Long id) throws Exception {
+        Event event = this.find(id);
+        if (!event.hasRegistrations()) {     // Controlo que no halla inscriptos
+            eventRepository.deleteById(id);
+        }else{
+            throw new Exception("No se puede eliminar el evento porque posee inscripciones");
+        }
     }
 
     @Override
     public List<Event> findEventsByOwnerId(long ownerId) {
         return eventRepository.findEventsByOwnerId(ownerId);
     }
-    
+
 }
